📝 docs(README.md): add instructions to run the upgrade process using 'php artisan pkt:upgrade' command
The README file was updated to include instructions on how to run the upgrade process for the package. The command 'php artisan pkt:upgrade' needs to be executed to perform the upgrade. This information was added to provide clarity and guidance to users on how to use the package effectively.